Relocate pipelinectl app files
