+++ internal/app/pipelinectl/cli/commands/drain/disable.go
