+++ internal/app/pipelinectl/cli/commands/drain/status.go
